+++ src/bin/ngit/sub_commands/repo/accept.rs
@@ -172,6 +172,10 @@ async fn accept_with_grasp_servers(
         .map(|lr| lr.description.clone())
         .unwrap_or_default();
     let web = latest.as_ref().map(|lr| lr.web.clone()).unwrap_or_default();
+    let upstream = latest
+        .as_ref()
+        .map(|lr| lr.upstream.clone())
+        .unwrap_or_default();
     let hashtags = latest
         .as_ref()
         .map(|lr| lr.hashtags.clone())
@@ -198,6 +202,7 @@ async fn accept_with_grasp_servers(
         root_commit,
         git_server: git_servers,
         web,
+        upstream,
         relays: relays.clone(),
         blossoms,
         hashtags,